Rechecked WIN dependencies

+++ src/start.c
@@ -42,7 +42,7 @@ void start(short port, const char* path)
 
     printf("\033[33m");
     // announce PID
-    printf("Starting Process: %i\tParent: %i\n", pid, ppid);
+    printf("Starting Process: %lli\tParent: %lli\n", (long long)pid, (long long)ppid);
     // announce root
     printf("Using %s for server's root", root);
     printf("\033[39m\n");
@@ -76,7 +76,7 @@ void start(short port, const char* path)
     socklen_t addrlen = sizeof(addr);
     if (getsockname(sfd, (struct sockaddr*) &addr, &addrlen) == -1)    stop();
     printf("\033[32m");
-    printf("Started Process: %i\n", pid);
+    printf("Started Process: %lli\n", (long long)pid);
     printf("Listening on port %i", ntohs(addr.sin_port));
     printf("\033[39m\n");
 }
